# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "EHPS API",
    "description" : "Equipment Health & Performance System — semiconductor fab equipment maintenance. Covers fab-coordinator machine/technician management, routine checkups, alert escalation and approval, repair assignment/completion, and the equipment health dashboard. All endpoints (except /api/auth/**) require a bearer JWT obtained from POST /api/auth/login, and are scoped by role: FAB_COORDINATOR, MANAGER, or TECHNICIAN. See docs/API-CONTRACT.md for the full endpoint contract.",
    "version" : "2.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/api/fab/technicians/{technicianId}" : {
      "get" : {
        "tags" : [ "Fab — Technicians" ],
        "summary" : "Get a technician by ID",
        "description" : "Returns a single technician managed by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "getTechnicianById",
        "parameters" : [ {
          "name" : "technicianId",
          "in" : "path",
          "description" : "Technician ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Fab — Technicians" ],
        "summary" : "Update a technician",
        "description" : "Updates an existing technician managed by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "updateTechnician",
        "parameters" : [ {
          "name" : "technicianId",
          "in" : "path",
          "description" : "Technician ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTechnicianRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/fab/technicians" : {
      "get" : {
        "tags" : [ "Fab — Technicians" ],
        "summary" : "List all technicians",
        "description" : "Returns all technicians managed by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "getAllTechnicians",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Fab — Technicians" ],
        "summary" : "Register a new technician",
        "description" : "Creates a new technician account managed by the calling fab coordinator. Requires role FAB_COORDINATOR.",
        "operationId" : "addTechnician",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddTechnicianRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/manager/technicians" : {
      "get" : {
        "tags" : [ "Manager — Machines & Technicians" ],
        "summary" : "List all technicians",
        "description" : "Returns every technician across the fab. Requires role MANAGER.",
        "operationId" : "getAllTechnicians_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UpdateTechnicianRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "email" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9._%+-]+@ehps\\.com$"
          },
          "phone" : {
            "type" : "string",
            "pattern" : "^\\d{10}$"
          },
          "speciality" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "name", "speciality" ]
      },
      "ApiResponseTechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/TechnicianResponse"
          }
        }
      },
      "TechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "speciality" : {
            "type" : "string"
          }
        }
      },
...
      "AddTechnicianRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "email" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9._%+-]+@ehps\\.com$"
          },
          "phone" : {
            "type" : "string",
            "pattern" : "^\\d{10}$"
          },
          "password" : {
            "type" : "string",
            "pattern" : "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
          },
          "speciality" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "name", "speciality" ]
      },
...
      "ApiResponseListTechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TechnicianResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}